[snes/spc] Add abstract reporting facilities
This allows the library to created without implementation details on how the reports
are graphically displayed (unlike before with STD).

In order to use, the function pointers must be defined in your higher level app. That is all!! Then the higher level app must learn to use these reports / collect them into a graphical display.