Locally built libraries can be used directly
Now that gme_m is a submodule of snestracker, it's important that the libraries
can be used directly without installing to any prefix. This fix will make sure a
non-versioned lib file is symlinked locally,

the addition of install_name is specific for Mac, I will have to check if it
works on Linux. It ensures the library is given an absolute path reference, otherwise
when binaries link to it, they won't be able to find the library since all they
had normally was the filename (no path).