[Make] Facilitate better autobuild / versioning
The versioning was just something I've been thinking about. I'd like to
try the versioning frame that goes Major . API-change . Minor. We'll see
how it goes. This library shouldn't change much afterall.

The subtle changes to the build process allow me to rely on the fact
that despite the version info in the library filename, it also installs
a general filename that snestracker build system can detect in order to
help it automate the building and installation of this library as a git
submodule